+++ sast-java/src/main/java/com/sast/astbenchmark/case_language_maturity/accuracy/context_sensitive/multi_invoke/need_solve/config.json
@@ -1,0 +1,14 @@
+{
+  "multi_invoke_need_solve": [
+    {
+      "evaluation item": "准确度->上下文敏感分析->多次调用->可求解",
+      "level": "4",
+      "scene_list": [
+        {
+          "compose": "DifferentParamsForFunction_001_T.java && !DifferentParamsForFunction_002_F.java",
+          "scene": "相同函数调用不同参数-scene1-可求解"
+        }
+      ]
+    }
+  ]
+}
