+++ sast-java/src/main/java/com/sast/astbenchmark/case_language_maturity/accuracy/context_sensitive/multi_invoke/without_solve/DifferentParamsForFunction_003_T.java
@@ -1,8 +1,6 @@
-package com.sast.astbenchmark.case_language_maturity.accuracy.context_sensitive.multi_invoke;
+package com.sast.astbenchmark.case_language_maturity.accuracy.context_sensitive.multi_invoke.without_solve;
 
 import com.sast.astbenchmark.model.alias.Invoke;
-import com.sast.astbenchmark.model.alias.X;
-import com.sast.astbenchmark.model.alias.Y;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -12,31 +10,25 @@ import java.util.HashMap;
 import java.util.Map;
 
 /**
- * Introduction 准确度->上下文敏感分析->多次调用->HeapAllocSite
+ * Introduction 准确度->上下文敏感分析->多次调用->相同函数调用不同参数-scene2
  * Level 2
  * Date 2024-07-05
  */
 // evaluation information start
 // real case = true
 // evaluation item =  准确度->上下文敏感分析->多次调用
-// bind_url = accuracy/context_sensitive/multi_invoke/DifferentParamsForFunction_005_T
+// bind_url = accuracy/context_sensitive/multi_invoke/without_solve/DifferentParamsForFunction_003_T
 // evaluation information end
 @RestController
-@RequestMapping("accuracy/context_sensitive/multi_invoke")
-public class DifferentParamsForFunction_005_T {
-    @PostMapping(value = "DifferentParamsForFunction_005_T")
+@RequestMapping("accuracy/context_sensitive/multi_invoke/without_solve")
+public class DifferentParamsForFunction_003_T {
+    @PostMapping(value = "DifferentParamsForFunction_003_T")
     public Map<String, Object> testcase(@RequestParam String cmd) {
         Map<String, Object> modelMap = new HashMap<>();
         try {
-            X x1 = new X();
-            X x2 = new X();
-            Y y1 = new Y();
-            Y y2 = new Y();
-            y1.f = cmd;
-            y2.f = "foo";
-            X rx1 = Invoke.assign(x1, y1);
-            X rx2 = Invoke.assign(x2, y2);
-            Runtime.getRuntime().exec(rx1.y.f);
+            String c1 = Invoke.id(cmd);
+            String c2 = Invoke.id("foo");
+            Runtime.getRuntime().exec(c1);
             modelMap.put("status", "success");
         } catch (Exception e) {
             modelMap.put("status", "error");
