+++ sast-java/src/main/java/com/sast/astbenchmark/case_language_maturity/accuracy/context_sensitive/multi_invoke/without_solve/MultiCallSite_001_T.java
@@ -1,4 +1,4 @@
-package com.sast.astbenchmark.case_language_maturity.accuracy.context_sensitive.multi_invoke;
+package com.sast.astbenchmark.case_language_maturity.accuracy.context_sensitive.multi_invoke.without_solve;
 
 import com.sast.astbenchmark.model.alias.Invoke;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -10,24 +10,24 @@ import java.util.HashMap;
 import java.util.Map;
 
 /**
- * Introduction 准确度->上下文敏感分析->多次调用->相同函数调用不同参数-scene2
+ * Introduction 准确度->上下文敏感分析->多次调用->MultipleCallSite
  * Level 2
  * Date 2024-07-05
  */
 // evaluation information start
-// real case = false
+// real case = true
 // evaluation item =  准确度->上下文敏感分析->多次调用
-// bind_url = accuracy/context_sensitive/multi_invoke/DifferentParamsForFunction_004_F
+// bind_url = accuracy/context_sensitive/multi_invoke/without_solve/MultiCallSite_001_T
 // evaluation information end
 @RestController
-@RequestMapping("accuracy/context_sensitive/multi_invoke")
-public class DifferentParamsForFunction_004_F {
-    @PostMapping(value = "DifferentParamsForFunction_004_F")
-    public Map<String, Object> testcase(@RequestParam String cmd) {
+@RequestMapping("accuracy/context_sensitive/multi_invoke/without_solve")
+public class MultiCallSite_001_T {
+    @PostMapping(value = "MultiCallSite_001_T")
+    public Map<String, Object> testcase(@RequestParam String cmd1) {
         Map<String, Object> modelMap = new HashMap<>();
         try {
-            String c1 = Invoke.id(cmd);
-            String c2 = Invoke.id("foo");
+            String c1 = Invoke.id8(cmd1);
+            String c2 = Invoke.id8(c1);
             Runtime.getRuntime().exec(c2);
             modelMap.put("status", "success");
         } catch (Exception e) {
