+++ sast-java/src/main/java/com/sast/astbenchmark/case_language_maturity/accuracy/context_sensitive/multi_invoke/without_solve/MultiCallSite_002_F.java
@@ -1,4 +1,4 @@
-package com.sast.astbenchmark.case_language_maturity.accuracy.context_sensitive.multi_invoke;
+package com.sast.astbenchmark.case_language_maturity.accuracy.context_sensitive.multi_invoke.without_solve;
 
 import com.sast.astbenchmark.model.alias.Invoke;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -15,19 +15,19 @@ import java.util.Map;
  * Date 2024-07-05
  */
 // evaluation information start
-// real case = true
+// real case = false
 // evaluation item =  准确度->上下文敏感分析->多次调用
-// bind_url = accuracy/context_sensitive/multi_invoke/MultiCallSite_001_T
+// bind_url = accuracy/context_sensitive/multi_invoke/without_solve/MultiCallSite_002_F
 // evaluation information end
 @RestController
-@RequestMapping("accuracy/context_sensitive/multi_invoke")
-public class MultiCallSite_001_T {
-    @PostMapping(value = "MultiCallSite_001_T")
+@RequestMapping("accuracy/context_sensitive/multi_invoke/without_solve")
+public class MultiCallSite_002_F {
+    @PostMapping(value = "MultiCallSite_002_F")
     public Map<String, Object> testcase(@RequestParam String cmd1) {
         Map<String, Object> modelMap = new HashMap<>();
         try {
             String c1 = Invoke.id8(cmd1);
-            String c2 = Invoke.id8(c1);
+            String c2 = Invoke.id8("foo");
             Runtime.getRuntime().exec(c2);
             modelMap.put("status", "success");
         } catch (Exception e) {
