+++ sast-java/src/main/java/com/sast/astbenchmark/case_language_maturity/accuracy/context_sensitive/multi_invoke/without_solve/config.json
@@ -5,10 +5,6 @@
       "level": "2",
       "scene_list": [
         {
-          "compose": "DifferentParamsForFunction_001_T.java && !DifferentParamsForFunction_002_F.java",
-          "scene": "相同函数调用不同参数-scene1"
-        },
-        {
           "compose": "DifferentParamsForFunction_003_T.java && !DifferentParamsForFunction_004_F.java",
           "scene": "相同函数调用不同参数-scene2"
         },
