--- sast-java/src/main/java/com/sast/astbenchmark/case_language_maturity/accuracy/flow_sensitive/loop_stmt/config.json
@@ -1,18 +1,0 @@
-{
-  "loop_stmt": [
-    {
-      "evaluation item": "准确度->流敏感分析->循环顺序执行语句",
-      "level": "2",
-      "scene_list": [
-        {
-          "compose": "Statement_ForStatement_005_T.java && !Statement_ForStatement_006_F.java",
-          "scene": "for-init流敏感"
-        },
-        {
-          "compose": "Statement_ForStatement_007_T.java && !Statement_ForStatement_008_F.java",
-          "scene": "for-update流敏感"
-        }
-      ]
-    }
-  ]
-}
