+++ sast-java/src/main/java/com/sast/astbenchmark/case_language_maturity/accuracy/flow_sensitive/loop_stmt/init/Statement_ForStatement_005_T.java
@@ -1,4 +1,4 @@
-package com.sast.astbenchmark.case_language_maturity.accuracy.flow_sensitive.loop_stmt;
+package com.sast.astbenchmark.case_language_maturity.accuracy.flow_sensitive.loop_stmt.init;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -9,26 +9,26 @@ import java.util.HashMap;
 import java.util.Map;
 
 /**
- * Introduction 准确度->流敏感分析->循环顺序执行语句->for-update流敏感
+ * Introduction 准确度->流敏感分析->循环顺序执行语句->for-init流敏感
  * Level 2
  * Date 2024-11-10
  */
 // evaluation information start
-// real case = false
-// evaluation item = 准确度->流敏感分析->循环顺序执行语句
-// bind_url = accuracy/flow_sensitive/loop_stmt/Statement_ForStatement_008_F/{cmd}
+// real case = true
+// evaluation item = 准确度->流敏感分析->循环顺序执行语句->for-init流敏感
+// bind_url = accuracy/flow_sensitive/loop_stmt/init/Statement_ForStatement_005_T/{cmd}
 // evaluation information end
 @RestController()
-@RequestMapping("accuracy/flow_sensitive/loop_stmt")
-public class Statement_ForStatement_008_F {
-    @GetMapping("Statement_ForStatement_008_F/{cmd}")
+@RequestMapping("accuracy/flow_sensitive/loop_stmt/init")
+public class Statement_ForStatement_005_T {
+    @GetMapping("Statement_ForStatement_005_T/{cmd}")
     public Map<String, Object> testcase(@PathVariable String cmd) {
         Map<String, Object> modelMap = new HashMap<>();
         try {
             String a = "mkdir";
-            for (int i = 0; i < 1; a = cmd + "|") {
+            int i = 0;
+            for (a = cmd + "|"; i < 1; i++) {
                 Runtime.getRuntime().exec(a);
-                i++;
             }
             modelMap.put("status", "success");
         } catch (Exception e) {
