+++ sast-java/src/main/java/com/sast/astbenchmark/case_language_maturity/accuracy/flow_sensitive/loop_stmt/init/config.json
@@ -1,0 +1,14 @@
+{
+  "loop_stmt_init": [
+    {
+      "evaluation item": "准确度->流敏感分析->循环顺序执行语句->for-init流敏感",
+      "level": "2",
+      "scene_list": [
+        {
+          "compose": "Statement_ForStatement_005_T.java && !Statement_ForStatement_006_F.java",
+          "scene": "for-init流敏感"
+        }
+      ]
+    }
+  ]
+}
