+++ sast-java/src/main/java/com/sast/astbenchmark/case_language_maturity/accuracy/flow_sensitive/loop_stmt/update/config.json
@@ -1,0 +1,14 @@
+{
+  "loop_stmt_update": [
+    {
+      "evaluation item": "准确度->流敏感分析->循环顺序执行语句->for-update条件求解",
+      "level": "4",
+      "scene_list": [
+        {
+          "compose": "Statement_ForStatement_007_T.java && !Statement_ForStatement_008_F.java",
+          "scene": "for-update流敏感"
+        }
+      ]
+    }
+  ]
+}
