+++ sast-java/src/main/java/com/sast/astbenchmark/case_language_maturity/accuracy/path_sensitive/loop_conditional_stmt/solver/Statement_AssertStatement_006_F.java
@@ -10,7 +10,7 @@ import java.util.Map;
 
 /**
  * Introduction 准确度->路径敏感分析->条件语句、条件表达式和循环结构->能够对上下文条件进行求解，以区分不同执行路径的状态->assert语句-条件成立
- * Level 3
+ * Level 4
  * Date 2024-09-18
  */
 // evaluation information start
