+++ sast-java/src/main/java/com/sast/astbenchmark/case_language_maturity/completeness/single_app_tracing/control_flow/conditional_stmt/Statement_SwitchStatement_002_F.java
@@ -26,10 +26,9 @@ public class Statement_SwitchStatement_002_F {
     public Map<String, Object> aTaintCase016(@PathVariable String cmd, @PathVariable String type) {
         Map<String, Object> modelMap = new HashMap<>();
         try {
-            switch (type) {
-
+            switch (cmd) {
                 case "ls":
-                    CmdUtil.run("mkdir" + " " + cmd);
+                    CmdUtil.run("ls");
                     modelMap.put("status", "success");
                 default:
                     modelMap.put("status", "success");
